requirements.txt
asgiref==3.7.2
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
Django==4.2.1
django-colorfield==0.8.0
idna==3.10
networkx==3.4.2
Pillow==9.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
reportlab==4.4.1
requests==2.32.4
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.5.0
